server/requirements.txt
fastapi
uvicorn
pydantic
pandas
numpy
scikit-learn
faiss-cpu
openai
python-dotenv  

